rename AldaParseFile/AldaPlayFile to AldaParseBuffer/AldaPlayBuffer
It turns out that I had already renamed AldaPlayFile to AldaPlayBuffer, but
forgot to update the help text to reflect the change.

I like calling it a Buffer in the command instead of a File, because File
implies that you have to save the file and the file is what's getting passed to
Alda, when what we're really doing is piping all the text in the buffer into
Alda -- which is more convenient because you don't have to save the file first.